initial state aware implementation of tristate
the tristate switch as the name suggests has 3 states. with toggle we can achieve only 2 states. the goal is the have state one as everything unmute, state two as microphone mute and speakers unmute and the last state as everything muted. using linuxes input library we can monitor the key code on an input node in devfs. on the oneplus3/3t, 600 601 and 602 are keycodes on event1 for tristate. although this implementation is not complete and currently only toggle works, but it now sees all 3 states then toggles it. it is an improvement for the future when we reimplement the pulseaudio mute and unmute functions to set different states instead of toggling